agents/rag_agent/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Answer questions related to Vertex AI documentation using Vertex AI RAG Engine and enrich with Google Search as needed"
authors = [
    { name = "Mike Goodman" }
]
license = { text = "Apache License 2.0" }
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "python-dotenv",
    "google-auth",
    "requests",
    "google-adk",
    "google-cloud-aiplatform[adk,agent-engines]",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-mock",
    "scikit-learn",
    "pytest-cov",
    "pytest-asyncio",
    "agent-starter-pack",
]

[project.optional-dependencies]
lint = [
    "ruff",
    "mypy",
    "codespell",
    "types-pyyaml",
    "types-requests",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["rag"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"
